Use predefined ring buffer when stack is empty
When the stack is created, 3 values are supplied and stored in a ring
buffer. If the stack is popped while empty, the first value in the
ring buffer is used and the ring buffer pointer is incremented. The
ring buffer is filled with x, y, t.

This prevents the stack from ever being exhausted, preventing blank
images when repeating binary operators, and leads to some very
interesting visual effects.