Use threading for image saving
Raylib's image saving function is a bit slow and the ffi Image type
cannot be sent across thread boundaries. To parallelise the image
saving, it grabs the ImageColors from the Raylib image, and then in
the threads it converts to image::RgbImage and saves them as
PNGs. This seems to give a minor (but unmeasured) speed improvement