Use pipe instead of saving images when creating video
Saving images as PNGs was the slowest part of video saving. Now we
pipe the image data directly to ffmpeg's stdin, which turns out to be
much faster. Note this removes the parallelisation but even so it is
still significantly faster. It also means we can increase the default
number of frames to save to 1000 in config.toml