Unfinished move from raylib to SDL
The reason was that the raylib bindings are incorrect for audio,
specifically that the audio callback takes an immutable buffer rather
than the mutable one that is necessary. However, the results were not
good (the output did not look as interesting, it was slower, and fonts
were not antialiased). It turns out on reading the raylib source that
raylib uses OpenGL to draw the rectangles, which I don't want to
bother implementing.