Fix colours in release mode
Before, in release mode, the colours looked wrong compared to the
debug build. This is because the `Color::color_from_hsv` function in
the Raylib C library casts to u8, but in C this is UB. It turns out
that in the debug build it was modding the value before casting to u8
but in release it was doing something else. The function is now
reimplemented in Rust with the wrapping behaviour explicitly
added. This was previously already in the sdl_rewrite branch